增加数据的权限控制

+++ src/main/java/com/ngtesting/platform/service/impl/CustomFieldOptionServiceImpl.java
@@ -1,6 +1,8 @@
 package com.ngtesting.platform.service.impl;
 
+import com.ngtesting.platform.dao.CustomFieldDao;
 import com.ngtesting.platform.dao.CustomFieldOptionDao;
+import com.ngtesting.platform.model.TstCustomField;
 import com.ngtesting.platform.model.TstCustomFieldOption;
 import com.ngtesting.platform.service.CustomFieldOptionService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,6 +14,8 @@ import java.util.List;
 public class CustomFieldOptionServiceImpl extends BaseServiceImpl implements CustomFieldOptionService {
     @Autowired
     CustomFieldOptionDao customFieldOptionDao;
+    @Autowired
+    CustomFieldDao customFieldDao;
 
     @Override
     public List<TstCustomFieldOption> listVos(Integer fieldId) {
@@ -20,7 +24,12 @@ public class CustomFieldOptionServiceImpl extends BaseServiceImpl implements Cus
     }
 
     @Override
-    public TstCustomFieldOption save(TstCustomFieldOption vo) {
+    public TstCustomFieldOption save(TstCustomFieldOption vo, Integer orgId) {
+        TstCustomField field = customFieldDao.get(vo.getFieldId(), orgId);
+        if (field == null) {
+            return null;
+        }
+
         if (vo.getId() == null) {
             Integer maxOrder = customFieldOptionDao.getMaxOrder(vo.getFieldId());
             if (maxOrder == null) {
@@ -37,14 +46,26 @@ public class CustomFieldOptionServiceImpl extends BaseServiceImpl implements Cus
     }
 
     @Override
-    public boolean delete(Integer id) {
+    public boolean delete(Integer id, Integer orgId) {
+        TstCustomFieldOption option = customFieldOptionDao.get(id);
+
+        TstCustomField field = customFieldDao.get(option.getFieldId(), orgId);
+        if (field == null) {
+            return false;
+        }
+
         customFieldOptionDao.delete(id);
         return true;
     }
 
     @Override
-    public boolean changeOrderPers(Integer id, String act, Integer fieldId) {
+    public boolean changeOrderPers(Integer id, String act, Integer fieldId, Integer orgId) {
         TstCustomFieldOption curr = customFieldOptionDao.get(id);
+        TstCustomField field = customFieldDao.get(curr.getFieldId(), orgId);
+        if (field == null) {
+            return false;
+        }
+
         TstCustomFieldOption neighbor = null;
         if ("up".equals(act)) {
             neighbor = customFieldOptionDao.getPrev(curr.getOrdr(), fieldId);
